Seperating LogCreator from LogWriter
