+++ MyChat.Tests/Report/IntegrationReportTest.cs
@@ -24,9 +24,9 @@ namespace MindLink.Recruitment.MyChat.Tests
             string[] args = { "--report" };
             var editorCofig = new EditorConfiguration(args);
             var editor = new ConversationEditor(editorCofig);
-            var writer = new LogWriter(editorCofig);
+            var logCreator = new LogCreator(editorCofig);
 
-            exporter.ExportConversation("chat.txt", "chatReport.json", editor, writer);
+            exporter.ExportConversation("chat.txt", "chatReport.json", editor, logCreator);
 
             var serializedConversation = new StreamReader(new FileStream("chatReport.json", FileMode.Open)).ReadToEnd();
